DDP-4276 remove copying feature from instance-creation action
