+++ pepper-apis/src/main/resources/org/broadinstitute/ddp/db/dao/EventDao.sql.stg
@@ -34,7 +34,6 @@ SELECT
      unpdf.pdf_document_configuration_id as user_notification_document_configuration_id,
      unpdf.generate_if_missing,
      aica.study_activity_id as activity_instance_creation_study_activity_id,
-     aica.copy_configuration_id as activity_instance_creation_copy_configuration_id,
      cp_answer.copy_configuration_id as copy_action_copy_configuration_id,
      ciea.contact_email_question_stable_code_id as contact_email_question_stable_code_id,
      (select stable_id from question_stable_code
