+++ study-builder/studies/study-schema.conf
@@ -552,11 +552,7 @@
         {
             "action": {
                 "type": "ACTIVITY_INSTANCE_CREATION",
-                "activityCode": "string",
-                "copyConfigPairs": [
-                    # List of copy pairings to perform after instance is created.
-                    # Optional, see `COPY_ANSWER` action for full schema.
-                ]
+                "activityCode": "string"
             },
             "dispatchToHousekeeping": "false",
             # trigger, precondition, etc ...
@@ -567,6 +563,9 @@
             "action": {
                 "type": "COPY_ANSWER",
                 # List of copy configuration source/target pairs, each specifying where to copy data from/to.
+                #
+                # If the trigger is `ACTIVITY_STATUS`, the triggering activity instance will be used as the source or target for copying.
+                #
                 # Current limitations:
                 # - `source` location is limited to `ANSWER` type.
                 # - If both source/target is `ANSWER` type, then question type must match.
